requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.13.1,<3.14
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
django-ckeditor==6.5.1
Pillow==9.4.0
celery==5.2.0
django-celery-beat==2.2.0
redis==3.5.3
django_celery_results==2.2.0
uwsgi>=2.0.20,<2.1